Updated Estimation and SheetItem entities and added migration
Updated `ApplicationDbContext.cs` to include new namespace `AppPlatform.Core.Enteties.EstimationEnteties` and defined new relationships for `Estimation` and `SheetItem` entities. Replaced the `Estimation` class in `Estimation.cs` with a new version including additional properties. Updated `ApplicationDbContextModelSnapshot.cs` to reflect these changes. Added a new `SheetItem.cs` file defining the `SheetItem` class. Added a new migration file `20240604204237_EstimationEnteties.cs` for creating and dropping the `Estimation` and `SheetItem` tables. The code changes also involve the addition of several classes and their properties in the `AppPlatform.Server.Migrations` namespace. These classes represent different entities in the application and are part of a migration named `20240604204237_EstimationEnteties`. The `EstimationEnteties` class is decorated with the `DbContext` and `Migration` attributes, indicating that it's a part of the Entity Framework Core migration. The `BuildTargetModel` method is overridden to define the schema of the database. This includes the creation of tables, defining their columns, setting up relationships between tables, and setting up indexes.