Added BidconBrowserModule and updated interfaces and project files
This commit includes the creation of a new project `AppPlatform.BidconBrowserModule` with its associated files and classes. The `IAccessClaimInfo` and `IApplicationLink` interfaces have been updated with new properties. Changes have been made to the solution and project files to include the new project. The `Routes.razor` and `Program.cs` files have been updated to include the assembly of `BidconBrowserModule`. New files have been added to the `AppPlatform.BidconBrowserModule` project, including a new image file and a JavaScript module.