Refactored layout with new PageSetupWithNavigation component
In this commit, we've made significant changes to improve code modularity and reusability. A new component, `PageSetupWithNavigation`, was introduced and used in `SettingsLayout.razor`. This component, defined in `PageSetupWithNavigation.razor`, takes two parameters: `NavigationContent` and `ChildContent`, and is responsible for rendering the navigation and main content of the page responsively. Additionally, `AppPlatform.Shared.csproj` was updated to exclude `components\PageSetupWithNavigation.razor` from certain build checks and include it as a non-compiled file.