Updated UI and added viewport resize handling
In this commit, several significant changes have been made. In `BrowserPage.razor`, new dependencies were added, `IAsyncDisposable` was implemented, and `IStringLocalizer<BrowserPage>` and `IBrowserViewportService` were injected. The UI was updated with new buttons and tabs, and logic for handling tab changes and viewport resizing was added. The `OnAfterRenderAsync` method was overridden to subscribe to viewport changes, and the `DisposeAsync` method was added to unsubscribe from viewport changes when the component is disposed.

In `_Imports.razor`, the `MudBlazor` and `AppPlatform.Shared.Components` namespaces were added to the imports. In `AppPlatform.Shared.csproj`, the `wwwroot` folder was moved from one `ItemGroup` to another. No packages or references were added or removed.

In `PageSetupWithNavigation.razor`, a class attribute was added to the `MudNavMenu` component, but the value of the attribute is empty. In `AppPlatform.Server.csproj`, the `wwwroot\ResizeListener.js` file was added to the project and is set to be copied to the publish directory.

Finally, a new file `ResizeListener.js` was added. This file contains a script for handling window resize events. It defines a set of breakpoints and a list of callbacks to be invoked when the window size changes. The script adds a resize event listener to the window that invokes the appropriate callbacks based on the current window size. The script also provides methods for adding and removing callbacks.